refactor(cmd): split runRun god function into stages
- runRun becomes thin orchestrator (~74 LOC).
- Stages extracted: validateFlags, setupSignalHandling,
  loadManifestAndPipeline, buildStateStore, autoRecoverResumeInput,
  resolveOrGenerateRunID, buildExecutor, assembleExecutorOptions,
  runOnce, runContinuous, formatRecoveryError, printSummary.
- runResources struct + Close() bundles emitter/audit/debug-tracer
  cleanups so callers defer one Close instead of tracking each.
- New files cmd/wave/commands/run_stages.go (stage helpers) and
  run_dryrun.go (performDryRun) keep run.go under the size budget.
- Each stage <= 150 LOC.
- run.go: 953 -> 375 LOC.
- No behaviour change; CLI flags, signals, output formats, and
  exit codes are unchanged.

Closes #1499.
